notebook rendering smoothed
