+++ src/traceml/cli.py
@@ -48,7 +48,7 @@ def run_with_tracing(
     print(f"Log directory: {log_dir}")
 
 
-    tracker = TrackerManager(interval_sec=interval, notebook=notebook)
+    tracker = TrackerManager(interval_sec=interval, mode='cli')
 
     # --- Arguments for the target script ---
     original_argv = list(sys.argv)
