+++ src/traceml/loggers/stdout/layer_combined_stdout_logger.py
@@ -6,7 +6,7 @@ from typing import Dict, Any, Optional
 import shutil
 
 from traceml.loggers.stdout.base_stdout_logger import BaseStdoutLogger
-from .display_manager import LAYER_COMBINED_SUMMARY_LAYOUT_NAME
+from traceml.loggers.stdout.display.cli_display_manager import LAYER_COMBINED_SUMMARY_LAYOUT_NAME
 from traceml.utils.formatting import fmt_mem_new
